Cleaned assignmens file and add template page to each component
