+++ src/components/Profile/AssignmentForm/AssignmentMain.js
@@ -39,8 +39,9 @@ class AssignmentMain extends Component {
       <div className="main-form">
         <div className="assignment-main">
           <div className="assignment-header">
-            <h1 className="h1">Assignments</h1>
-            <button type="button" className="waves-effect waves-light btn orange">Add</button>
+            <button type="button" className="waves-effect waves-light btn orange">
+              Add
+            </button>
           </div>
           <table className="responsive-table mainTable">
             <thead className="assignment-main-header table-header">
@@ -56,7 +57,6 @@ class AssignmentMain extends Component {
             <tbody>{itemsform}</tbody>
           </table>
         </div>
-        <AddAssignmentForm onSubmit={this.submit} />
       </div>
     );
   }
